+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to a section with the version name.
 * Repeat invoice form lines.
 * Increase total amount font size on collection form.
 * Reports supports user permissions.
+* Filter customer parties to be customers/ pampers only.
 
 ## 1.2.0